Simple implementation for behavioral cloning
The script runs behavioral cloning for the cart-pole environment, and
the expert is obtained from the file cartpole_model, which contains the
TensorFlow variables from a previously trained network using OpenAI
baseline's DDQN.
The expert model is sampled and the samples are saved in a replay
buffer. The clone of our expert is just a neural network trained over
the samples provided by the replay buffer.