Add behavioral cloning for double pendulum
The expert was obtained from the DDPG algorithm in garage and is
retrieved by loading the corresponding TF session for the training of
the expert.
Also, the data sampled from the expert is stored in a numpy file to
avoid using a replay buffer, considering that the data set will be
fixed.
Finally, the script for discrete actions as updated to work with the new
sampler in garage.